arm64: dts: lx2160a-qds: make the QIXIS CPLD use the simple-mfd-i2c.c driver
The MDIO mux on the LX2160A-QDS never worked in mainline. There is a
downstream driver for the QIXIS FPGA which is very similar to the
already existing drivers/mfd/simple-mfd-i2c. Since the HW works with
simple-mfd-i2c.c there is no point in upstreaming the other one.

Adapt the compatible string and the child node of the FPGA node, so that
the simple-mfd-i2c.c driver accepts it.

Signed-off-by: Ioana Ciornei <ioana.ciornei@nxp.com>